requirements.txt
--extra-index-url https://alpine-wheels.github.io/index
backports.zoneinfo;python_version<"3.9"
daphne==4.0.0
channels==4.0.0
Django==4.1.7
psycopg2-binary==2.9.3
djangorestframework==3.14.0
pandas==1.5.3
numpy==1.24.2
openpyxl==3.1.1
django-cors-headers==3.14.0
celery
django-redis
six